+++ src/pipeline_components/content_optimizer.py
